Fixes issue #415: Clicking on a tab should not change its position when "Multiline tabs" is selected
The behavior was caused by an automatic (and erroneous) shift of tabs when changing active page in multiline tabs. The row with the selected page is moved near to the editor (behavior of the control) and this movement was treated at the same time as a drag and drop shift.
For now the movement of the row is maintained, although tabs in each row keep its order.

Signed-off-by: Daniel Prado Velasco <dpradov@users.noreply.github.com>