Included two new configuration options to control how internal links (KN Locations) are shown
CaretInKNTLinks  (1 or 0 (default))
RelativeKNTLinks (0 or 1 (default))

When RelativeKNTLinks=1 the path of the insertion point is compared with the path of KNT hyperlink. Common ancestors are hidden.
The caret position will only be included with the text shown if CaretInKNTLinks=1.

By now they are only modifiable from .INI file, in its section, TreeOptions.

(Relates to #528: Minor improvements in links)

Signed-off-by: Daniel Prado Velasco <dpradov@users.noreply.github.com>