Some corrections and cleanup in PerformCmd and PerformCmdEx
When a note is marked (or unmarked) as Readonly, the file is marked as modified but it is not shown in the status bar.

As a consequence of #533 (and the change in PerformCmd (adding ActiveNote.Editor.BeginUpdate / EndUpdate) was RxRTFChange was not called and that caused that the node (if executed in a tree node) wasn't marked as modified.
By other way, in RxRTFChange was only modified NoteFile.Modified and SelectedNode.Modified (in TreeNotes). ActiveNote.Modified is updated in kn_noteObj:EditorToDataStream (but only with TTreeNote...) or ...
>> It is essential to do a thorough refactoring. For this one and many other potential problems and the cost to make any change or find the origin of any bug.

Related to #536 : Various corrections and improvements

Signed-off-by: Daniel Prado Velasco <dpradov@users.noreply.github.com>