Adjustment to change in #520
With the change made in #520, in EditorToDataStream (TTabNote and TTreeNote), to only copy from Editor to Stream of node/note when there are modifications in the Editor, its necessary that the stream of the node/note is always available. But in TTabNote.DataStreamToEditor, after copying to the Editor the data stream was cleared. I think that it was a risky optimization, based on the fact that EditorToDataStream was always called at the end.

Without this correction, the changes in #520 would cause the RTF notes to be saved empty!

Signed-off-by: Daniel Prado Velasco <dpradov@users.noreply.github.com>