Adjusts issue #38 when line wrapped (#38: HOME key, could put caret on first non-space position)
When wrapping is enabled and we are at the final of one line SelStart could return the same value as if we were at the beginning of the next one (*). In that case, pressing HOME would move the cursor incorrectly to the first character of next line.

(*): That occurs when we go to that position pressing 'End'. If we move (left arrow) from the beginning of the next line, the cursor will be apparently in the same position, but in the first case it will be as if we were in the next line.

Signed-off-by: Daniel Prado Velasco <dpradov@users.noreply.github.com>