Fixes #565: Paste as text not working in Beta 8
"Paste as text" was doing nothing when adding (not when replacing a selection) more than one line.

It is related with changes in issue #564. Apart from the corrected error, this modification takes into account that in some incorrect replacements (by the paste as text operation) the selection is not modified, nor replaced with the plain text nor deleted, but simply unselected. It now distinguishes this case, because it should not emit any Undo operation, or it could lead to undo any previous and correct modification.

Signed-off-by: Daniel Prado Velasco <dpradov@users.noreply.github.com>