Adaptation from Borland Delphi 2006 to to Embarcadero Delphi CE 11.3 (Community Edition 11.3  Alexandria).
I have abandoned some 3rdParty components/libraries, incorporated new releases of some of them, and patched many of the others.
In the case of UnRxLib (which I have used to replace RxLibrary), I have made a series of changes to several files, specially
RxRichEd.pas. Although the changes are marked as such, I include in this commit only the original files, and in the next commit
I will apply the changes (tested with all the rest of project uploaded in this actual commit).

Along the way I have identified some errors present in actual version of KNT (and some minor improvements). Some of them have
been corrected during the process of adaptation to Delphi CE. But in others I have preferred not to correct them at this moment,
to not mixing them with all the changes necessary to this adaptation (many files, many changes...). I will make specific commits.

I have decided to drop support for the 'Dart format'. In the case that someone needs to reuse a file in that format, can use
an older version of KeyNote to convert to normal KeyNote format.

(More information in doc\README_SourceCode.txt and doc\README_News.txt)