Fixed: Ensure that all nodes in a RTF tree note are saved in RTF format,        and all nodes in a plain text only tree note are saved in plain format.
when dropping a file with plain text into a new node of a RTF Tree, or when dragging and dropping a node with plain content
(from a tree configured as such) onto a RTF tree, We must ensure that the node's stream is loaded with it's RTF
conversion, to avoid ending up saving that node's stream content in plain format to the .knt file when saving
(could be problematic when reading the .knt file).
Something similar is needed in the opposite case (RTF content into a plain tree), although in this case the only problematic
case would be to import a plain file dropping it into a new node of a plain tree (more information in the comments to the
changes).