Improvements on Clipboard Capture / Web Clip capture: ensure to include Title for URL
So far, when pasting with Clipboard Capture (and 'Include source URL' selected) or with Web Clip, from a web page,
only the URL was include, no the title. At the beginning some browsers included that information on the clipboard,
but currently it does not.
Now KeyNote NF will try to get the title using WebBrowser (offline, using cache).
Although it tries to process as little HTML text as possible to get it, the time it takes can depend on the page.
Normally it wil be very fast, but there are very heavy web pages on witch it can takes a little more time.
For this reason, the application includes now three tokens to referentiate source URL in the divider string:

  '%S'  insert Source URL (with title)
  '%s'  insert source URL (with title, limited waiting time)
  '%U'  insert source URL (without title)

  With '%S', it will get the title for the URL, without time limit (6 seconds, just for security)
  With '%s', If title is not figured out in less that 2 seconds (normally will be much faster) only URL will be used
  With '%U', KeyNote will only show URL, not title (even though it's already available (*) )

(*) A small cache will be used for titles, in case you need to paste several times, from the same pages.
So, even certain delays will only apply to the first capture on a certain URL.

There is more info on on Clip Web and Clipboard Capture on:
'Improvements on Clip Web and Clipboard Capture #532' (https://github.com/dpradov/keynote-nf/issues/532)