Title from YouTube page: exception to the normal paste behaviour
From now it seems not possible to obtain the title of a page of YouTube, because the html returned by TWebBrowser
in this case does'nt contain the final html page.
So, KNT will not try to guess title if URL begins with https://www.youtube.com
As a easy shortcut to create a hyperlink that includes the URL and the caption of the video, we can do simply:
- Select only the caption of the video (just click three times in one word of the video title)
- Copy Paste (or Web Clip)

If the text selected includes only 1 line and it hasn't more that 100 characters, then it wil be assumed as the title
and a normal hyperlink with Youtube URL and that title will be inserted.

(Note: https://www.smperth.com/resources/youtube/youtube-character-limit/ )