Fixed: copying and pasting from certain web pages can produce nothing or truncate some text
KeyNote will paste from web page as plain text (without any formatting) when detected that conversion from HTML to RTF it is no ok

It can occur that the conversion to RTF (obtained with TWebBrowser) from the HTML copied to the clipboard from certain pages
are totally o partially incorrect (even MSWord can't convert correctly that fragments).
I have detected this problem with the following site, whose html seems to be odd: stackoverflow.com. But it could happen
with other site/s.

For example, consider the following URLs and try to copy and paste several paragraphs altogether.
       https://stackoverflow.com/questions/46991729/how-to-load-a-stringstream-with-utf8-characters-to-a-stringlist-in-android
       https://stackoverflow.com/questions/11122197/why-does-memo-lines-use-tstrings-instead-of-tstringlist

Most of them will produce '<>' or '', others '<<<' or something similar and a few can truncate some text.
If in this examples we copy (and paste) only fragments of certain paragraphs or just snippets of code, it seems to be ok.