Pasting as plain text could move the caret several characters before its correct position
In solution to issue #539 (Pasting as plain text moves the caret beyond pasted text when it includes several lines)
this problem was partially mitigated and caused, on certain occasions, the opposite effect, placing the caret
ahead of the position that corresponds to it.

Now I have changed the way the plain text is pasted, and I think it has finally been correctly resolved.