Clipboard Capture / Web Clip. Added new token: '%%'
Added a new token (%%) that delimits a divider string for the second and following consecutive paste
from the same page (same URL).
KNT will remember the last URL it was pasted from in the current [Clipboard Capture] session.

This allows, for example, to include the URL and title only before the first fragment we paste from a certain web page. Next (consecutive) fragments of the same URL can include any of the tokens, but normally we will only need a simple delimiter (like '- - - ').

Predefined divider strings have beend updated, including this new token and allowing this behaviour.
It is of course possible to create a custom divider, and you don't have to use the new token in it.