Fixed: TColorBtn had a strage visual behaviour after migrating form Delphi 2006 to Delphi CE 11.3
If configured as Flat=True (normal and appropiate in this custom button) when hovering the cursor on the
button it was highlighted in blue, but not temporarily while the cursor was over, which would be normal.
That blue color was not removed, but 'added', until it was left with an intense blue. It was as to paint with
a blue brush on top.