At the moment, the Send Mail functionality will not be included in KNT
It would be necessary to update it to manage correctly authentication (username and password / smtpAuthXOAuth2 /  smtpAuthOAuthBearer), and I think that this is not, probably, a demanded feature of KeyNote.

Besides, with the update to 'ICS - Internet Component Suite - V8.70', by Francois Piette, it implies
almost 600kB extra in executable file's size.