New: Added 'KeyNote' as target format in File/Export...
Now it is easy to create new KeyNote files from current file, with the export mechanism:

- Including only active note
- Including only current node and subtree of the active note
- Including only checked nodes in active note
- Including selected notes
- Including all notes (this is equivalent to File/Copy To.., but with the possibility to select the file name)

In all that cases it is possible to consider only not hidden nodes. For example, if we run 'Find All' with
'Filter Tree note' option, we can then generate a new .knt with the nodes that include the searched pattern.

* The create .knt file will have the same properties as the original (native, compressed, ..)