Modification in RxRichEd to change the way a retry is made when the load of stream detects an error
It will retry with no TEncoding conversion, setting only SF_TEXT (no SF_UNICODE). The stream will be
supposed to be in UTF8 format, with BOM. RichEdit will manage it correctly.
This is how TRichEditStrings.LoadFromStream (in RxRichEd with KeyNote) worked in previous versions.
Now those symbols incorrrectly converted (with previous versions) wil be ignored, but the rest will be
shown correctly as UTF8.

(Resolves issue #610)