Fixed: Note Properties. Changes in tree font didn't update correctly
Font changes in the tree (from 'Note Properties...') were not reflected on it on many occasions,
precisely those in which we had not modified the background color.

It was also happening that the font changes, when they were shown, did not give rise, many times,
to a correct resizing of the height of each item.

* These problems were of tree refresh. Saving and reopening the file would show up correctly.