Fixed: (Important) Text files imported as a new note note could be saved plain, not RTF
This problem exists in KNT from the beginning, but with the changes included from version 1.8.0
could be much more problematic, when loading the .knt file

When importing a file with plain text into a new note, we must ensure that the note's stream is loaded with its RTF
conversion, to avoid ending up saving that node's stream content in plain format to the .knt file when saving.
This could occur if we didn't modify the text of that new note.

This was fixed when importing as nodes, but not applied also for files imported as notes (...!)
Was corrected for importing as nodes in Commit ac672bb258... 11/06/23:
 * Fixed: Ensure that all nodes in a RTF tree note are saved in RTF format,
   and all nodes in a plain text only tree note are saved in plain format.