New: pressing [Ctrl] + <Hide Tree Panel> will temporarily preserve editor width, reducing the application window width
When the user hides the Tree Panel pressing Ctrl (usually: Ctr + Shift+F11  or Ctrl + [View | Tree Panel] ) then,
instead of letting the editor panel grow to the left occuping the width of the tree panel, the application window will
reduce its width temporarily so that the editor maintains its width and position.

Once shown the Tree panel again, the application window's size and position will be restored.

If, after hiding the tree panel in this way (the window's width is now reduced), the user selects another note, then
the application will automatically make the tree panel visible and restore the aplication window width and position,
before changing to the other note.

Opening another file or closing the application while the windows's width is reduced, will restore it to it's correct
size and position.

This functionality can be disabled in keynote.ini, with a new option: IgnoreCtrHideTrePanel = 1