Fixed: New KNT Links could jump to an incorrect position (see description)
KNT links that point to a hidden mark look for a certain pattern and ID bookmark. When the application copies text to the
clipboard, it removes the hidden bookmarks. Not doing so could make that if we pasted the text above the one we were pointing to,
the link will find the hidden mark before the correct one. (On the contrary, the cutting operation does need that hidden mark.)

In normal situations, the hidden mark is maintained by the control RichEdit as we inserted it. But I have found that RichEdit
can alter some of them, adding control tags within the block that we have bounded between \v and \v0 (actually through {\v....} ).

(This is a problem only when we are looking for the hidden marks through RTF syntax. In most situations, KNT works directly with
hidden characters)

We may find that an added tag like:
	...    \v\''11B2\''12\v0 HELLO
it ends up appearing for example as:
	\v\f0\fs16\lang3082\''11B2\''12\v0 HELLO

We must convert:
  "\v\f0\fs16\lang3082\''11B2\''12\v0 HELLO"   ->  "\f0\fs16\lang3082 HELLO"