Added option to replace bullets by alternative string when pasting as plain text
By default, Windows replace bullets by a symbol in Cambria Math (something like a big point) plus a tab character
In RTF it is of the form:
\f1\u10625?\f0\tab One
\f1\u10625?\f0\tab Two

With the new option is possible to replace the symbol and tab characters by other text. By default it is set to "" (empty string)
but can be configured (in keynote.ini) with something like " - ", for example.

[EditorOptions]
...
BulletsInPlainText=" - "

Issue related: #612 don't copy bullets to other apps