Fixed: Searches may not correctly highlight results when images are visible (if storage <> EmbeddedRTF)
The positions calculated in the searches (Find All and Find Next) are based on the TextPlain obtained with the images hidden
(the format used when saving notes/nodes). When all images are hidden (imLink mode), or there are no images, these positions
allow matches to be highlighted directly, but when all or some of the images are visible in the editor it is necessary to
adapt these positions, calculating the offset to be considered at the time when it is required to jump to or highlight a match.