Fixed: Missing dot-separator on files saved in external image storage
The problem is described here:
https://github.com/dpradov/keynote-nf/issues/623#issuecomment-1819696740

However, the problem does not seem to have anything to do with what is pointed out.
<<And yet, enabling the Change button seems to have a different effect (no missing dot-separator) as opposed to leaving the Change-button alone (missing dot-separator), I just tried it again by converting a new file back and forth between embedded and external.>>

It's something simpler. It seems that from some locales the expression 'FormatDateTime( 'ddMMM', Now)' used as part of the file name, adds a dot to the end and in others it doesn't. In the spanish locale it adds the point and I supposed that it was something general.
Now the code does not depend on that (possible) dot to separate the extension. It constrols it explicitily