Fixed: Inserting high resolution images for 100% viewing does not add the images
As it is described in https://github.com/dpradov/keynote-nf/issues/623#issuecomment-1824896077
and https://github.com/dpradov/keynote-nf/issues/623#issuecomment-1824370437

It seems that the RichEdit control supports a maximum visible value, in any of the dimensions (x or y)
of approx. 32760 twips (2184 pixels) (1 pixel=15 twips).

Now KNT will consider that limit of 32760 twips. If it is exceeded in any dimension, it will be limited to that value, proportionally reducing the other.