* Fixed: Image could be duplicated on storage (with new ID) after pasting an image previously copied or cut to clipboard
The reason for this duplicity is explained, for example, here:  https://github.com/dpradov/keynote-nf/issues/623#issuecomment-1834392261
Despite the latest corrections made, this problem could still occur when pasting an already registered image from the clipboard,
previously copied or cut.

It was not controlled that the size of the RTF string constructed by the application from the image stream could be greater than
that available from the clipboard, because the RichEdit control will have eliminated the following information if it was present:
Embedded color profile, EXIF metadata, IPTC metadata

Among the consequences: the hidden mark was not being added, so it would end up being interpreted as another image. Besides, data
could be being copied beyond the reserved buffer (In debug mode a Range check error exception is generated...), which it could be
causing other errors that are more difficult to detect.