* Images. Added "Restore image[s] proportions" to the RTF context menu
As requested here: https://github.com/dpradov/keynote-nf/issues/623#issuecomment-1830661325

The new action will reset proportions and reconsider "Max.auto width" on selected image[s].
If there is no selection, it will modify all images in the editor.

The functionality of the Alt+Click (on toolbar image button) has been expanded to allow acting
only on the selection, if any.

As before, the resizing done with Alt-click, when it is applied to all images in the Editor
(with no selection) is temporary by default: if the node has no modifications, its status will
not change, it will continue to remain unmodified and these changes will not will be saved.
But if you want to keep them, just make any minimal change to the node (such as adding or
removing a space...), to mark the node (or note) as modified.
But when applied to a selection, the changes will mark the node (or note) as modified. In this
case, the changes in the images can be reverted with Undo operation.

The new action, "Restore image[s] proportions", by the opposite, will always mark the node/note
as modified, even when applied to all images in the editor. Changes can also be reverted with
Undo operation.