Fixed: Can be difficult to open images inside tables with the internal or an external viewer
Problem described here: https://github.com/dpradov/keynote-nf/issues/623#issuecomment-1845791606

It is possible to include images in the cells of a RTF table, but due to the behavior of the RichEdit control, it could be difficult to open the image (by double clicking, or Ctr+double clicking) because in many cases not only the image is being selected but other characters in the same cell or row of the board.
It was possible to open it, trying again, but difficult. In hidden images mode (seeing the link to the image) this problem does not occur and has always worked correctly.

The application is modified to identify and address this situation. Images contained in tables can now be opened normally.

Important:
Images contained in table cells cannot be resized manually, the RichEdit control does not allow it (can be checked with WordPad). Instead it will try to perform a drag and drop action. If the image is moved in this way, you will have to reposition it by cutting and pasting or dragging it again, as you will not be able to Undo.
It is possible to modify the dimensions using the 'Restore image[s] proportions' option or by Alt+Click on the image button, while selecting this image. If you need to set a size at will and not depend on the 'Max auto. width on insert', perhaps the simplest thing is to cut and paste the image outside the table, resize it and then reposition it in the table.