Images. New: Ctr+Click on View Images (menu and toolbar button) -> Reapply hidden if Images Mode is imLink (images hidden)
Until now, clicking with CTRL on the button forced the reload of the images in the editor, to allow reconsider from disk
those images edited externally (having already been uploaded/cached).
This function now applies only if Images Mode is imImage (images are displayed). If Images Mode is imLink (images are not
displayed) it will now force any images that may be visible to be hidden.

This is useful if you are inserting many images and need to hide them from time to time. Even if the Mode is imLink,
added images are always shown.
It is now possible to hide these new images without having to previously show all the others.