Manage the interruption of Copy and Cut operations from KNT when ClipCap is active
We must call LogRTFHandleInClipboard from the very beginning of ClipCap detection, when internal ecCut or ecCopy operations have been interrupted
just after sending content to the clipboard, before they can do the call. We are using that handle to identify in several points if the Clipboard
content was copied by the application itself
It does not seem to have any impact on the ClipCap operation, beyond perhaps preventing some optimization of image processing.
But it seems coherent to ensure that it is called.