Export. Some minor adjustments related to note/node headings
Symbols will only be applied to node headings. It seems more normal, and makes clearer the use of symbols options.

Also, when adding the token %>, the length calculated from "Length heading" is now reduced if option 'Indent nodes' is checked.
This reduction depends on node level and the value set in "Tab size" (in the Note properties). When exporting to plain text,
it also depends on the relation between "IndentValue" and "Increment indents by", which defines the number of 'tabs' for each level.

Sceenshots to show the difference in https://github.com/dpradov/keynote-nf/issues/573#issuecomment-1872048118