Glossary: shortcuts formed exclusively by non alphanumeric characters will be automatically recognized
We can expand "--" to "—", for example, using F7. It is not necessary to select previously those twp characters.
KeyNote will first try to find a term in glossary considering spaces (' ', tab and break lines) as word delimiters.
If it does not find any terms, it will repeat the search considering any non-alphanumeric character as a word delimiter (as usual)

So, if we have, "--" -> "—"   and "eg" -> "Example", then:
"--eg" --> "--Example"
If we also had "--eg" -> "EXAMPLE", then
"--eg" --> "--EXAMPLE"
"-- eg" --> "-- Example"

We can define something like "*10" -> "*********"