Improved Bookmarks 0-9: Now are persisted, and are linked to target through a hidden mark
Using the same mechanism that Keynote's Mark/Insert Location is based on, we can now use the 9 auxiliary markers:
- Search|Set Bookmark (Shift+Alt+[0-9])
- Search|Jump to Bookmark (Alt+[0-9])

As before, the markers are global to the entire knt file.
These markers are saved within the file, as well as the small hidden markers used as target.

Tip: When using the keyboard shortcut Shift+Alt+[0-9] it may be convenient to use the ring and index fingers
of the left hand (4 and 2, according to piano terminology), although it is also possible to assign other shortcuts.