New: Added "Check for Updates" to Help menu. Can also be checked at startup.
The dialog box "Check for Updates" offers the option "Check for updates on startup" (KeyOptions.CheckUpdOnStartup) (Default: True (1))
It compares your version (.exe running) and current version (in Github, looking in releases/latest)
showing the changes between those versions based on history.txt file

If "Check for updates on startup" is enabled, it will be checked several seconds after initialize KeyNote, and only once a day.
When found a new version, and shown the dialog box, it will register that version as the last informed version, to only notify
again if there is a newer version.

* Another change:
  KeyOptions.TipOfTheDay won't be automatically enabled when detected a new version (shown if KeyOptions.IgnoreUpgrades is disabled)