* Improved Tools | Customize Keyboard...
Now it is possible to set any shortcut to Macros, Plugins, Templates, Styles and Fonts.
The shortcuts assigned to menus (main or tree) will take priority over shortctus to other commands (macros, ...)
The design and usability have also been reviewed.

Funckey plugin it is not necessary now, and it is abandoned. That plugin required KeyNote to listen to a certain
message after change was done, to reload shortcuts, and to use a specific config file (keyboard.key), with a limited set of possibles shortcuts.

This closes issue #30 and partially cover what it is requested in issue #645