Changes in behavior of 'View Resource Panel' shortcut
If you press the 'View Resource Panel' shortcut (default F9), the focus will shift from the Editor or Tree panel
to the resources panel if it is visible. If the resource panel had focus, it will be hidden.
The rest of the behavior is the same as current: if the resource panel is hidden, the shortcut will make it visible.
Double-click the status bar or click the menu entry (View|Resource Panel) to hide or show the Panel resource, as usual.