New optional behavior of Tree panel
If width of tree panel is changed with Ctrl pressed, the new value will be considered the max. width,
with the following behavior:
- When you mouse over the tree panel, at any node, the tree panel will expand to the maximum width.
- The same happens if the focus is moved to the tree panel.
- Once focus returns to the Editor panel, the tree restores its normal (minimum) width

- You can remove the max.value clicking on the splitter with Ctrl pressed, but without changing the width.
- You can also keep fixed the current width (normal or maximum) pressing Alt while clicking on the splitter.
  This is highlighted changing the color of splitter to light blue. This way the tree width will not changed
  automatically.
  Clicking again with Alt pressed will deactivate this fixed state.
  If in this state you change the width of the tree panel, it will also deactivate the fixed state.

Both values, normal width and optional max. width, will be saved on .knt file

(In case of vertical layout of Tree Panel, Heigth instead of width will be used)

* A new INI option has been added:

 ModifiedOnTreeResized  (default 0, False)

If set to 1 (true), changing the width (or max width) of any tree panel will mark the file as modified, and so
this changes can be saved.