Fixed: When playing macros of certain commands, all of them were executed as the first one
When reproducing the macros, a string list type variable is used to separate the command itself from its arguments, on each line. For example, a line can be "#insertchar|148|1|Tahoma|0", and must be separated into the command (insertchar) and its arguments (148, 1, Tahoma, 0).
Well, that object was not being cleaned in each iteration (each line of the macro), so it must have been causing four commands to work poorly (those that use arguments), in which each execution of one of those commands would be reproducing really the first of them. Commands affected:

ecFontDlg
ecParaDlg
ecInsCharacter
ecFindText

Ref: https://github.com/dpradov/keynote-nf/issues/646