Changes in reloading of resource panel data
Resource panel content is always loaded only when necessary.
The option "Load Resource panel data only when necessary" (in Options|Advanced) has been renamed
to "Reload Resource panel data when displaying" (ResPanelActiveUpdate), because that is it's function.

Aside from that change in the description, the way the update is done has also been changed.
Instead of always clearing the respanel content when hiding it and reloading it when it is shown again, the content
is updated (by clearing and reloading) just when the Resource panel is displayed again.
This way the macro list is not removed when the resource panel is hidden, preventing you from using shortcuts that run macros.