Fixed: When restoring image(s) proportions and the Note is Read Only, always use a temporary method
If the note is marked as Read only, the temporary method will always be applied, regardless of whether you use Alt+Click
(with or without selected text) or the "Restore image[s] proportions" command. This also means that all images in the
node will be adjusted, and not just selected images.