New option, "Show hidden marks and IDs", when exporting to plain text
The additional "Show hidden marks and IDs" option makes the hidden marks used with internal KNT Links, bookmarks and images visible.
It also makes hidden part of hyperlinks visible and shows the internal IDs of notes and nodes.
This can be useful if you are creating a help file and you want to open the file externally showing a specific location.

The hidden marks are shown in a humanized way. Example:

$11B7$12  	-> [BMK:7]
$11b7$12  	-> [TmpBMK:7]         (Bookmarks 0 - 9)
$11I7$12	-> [IMG:7]

The IDs of notes and nodes is shown enclosed in "[" and "]". Example:
My NOTE [1]
  My node B [123]

The hyperlinks are visible in the following way (eg. KNT link):

HYPERLINK "file:///*8|282|806|0|11"Text of hyperlink

* Also, %I token (Node index) now begins with 1