* Added new command line option: -title"<TITLE>" Title to use in main window (mainly for its use with kntLauncher)
If passed to kntLauncher (recommended), this add-on will search for an existing KeyNote instance with the file
to open looking for the title of the window (and also the classname of KeyNote app). In this case it will use
FindWindow and not EnumWindows.

* Also: Corrected the processing of command line received from KNT_MSG_LAUNCHER_CALL or KNT_MSG_OTHER_INSTANCE_CALL
messages (was incorrectly split in its argument)