Improvements in "Normal vs maximum tree width" functionality
The new behavior is:
- When you mouse over the tree panel, at any node, the tree panel will expand to the maximum width.
- The same happens if the focus is moved to the tree panel.
- If the Editor panel maintains the focus and the tree panel is expanded, if you mouse over the editor then the tree restores its normal (minimum) width.
- Also, when focus returns to the Editor panel, the tree restores its normal width

Tip:  You can maintain the tree panel expanded by keeping the mouse over the tree panel while you type in the editor at the same time.

Additionally, it is now possible to traverse the tree panel with the mouse and not trigger tree width expansion (just move it a little faster).
Besides, it has been avoided the selection of text that sometimes occured as a side effect of switching the focus from tree panel to editor by
clicking on the editor panel (restoring the tree panel width).