Fixed: Favorites in Resource Panel need to be refreshed in certain ocassions
I was finally able to reproduce this issue once, enough to identify and fix it

With the help of the 3rdParty TFormStorage control (RxPlacement.pas) KeyNote recovers the position and state of the windows (saved in the keynote.mru file). This occurs when the main KeyNote window is displayed, before it is activated.
From what I had been observing, it seemed that this process ( =>TFormPlacement.RestorePlacement) ended before continuing to process other messages. But it didn't have to be that way.
Sometimes (in my case very rare occasions) the Activate message (on the main form) is received before that process has finished. In this case, the active tab in the resources panel may not have been restored yet. But nevertheless, it is in FormActivate where the initial loading of the resource panel is carried out (on its active tab)
Fortunately the TFormStorage control offers an event that reports the completion of the restore placement process.

(In my case it is something very, very unusual. Few minutes after adding the code that clearly corrected the problem, the favorites panel began showing correctly on the startup, even commenting the new code)

Refs:
 Favorites in Resource Panel need to be refreshed in 1.8.xx #631
 "Resource Panel>Favorites, not autocomplete" #655