Fixed: Glossary doesn't display correctly items that include '=' character inside the shortcut
The '=" is used to separate the Name from the Value (Name = Value). To allow seamless use of that character in the name KeyNote
now saves that escaped character with the '\' character.

For example:
=> expands to ⇒

will be saved as:
\=>=⇒

Once loaded the file, KeyNote will replace internally the '=' separator with another non problematic character

Ref: Some issues with the Glossary #671