New: It is possible to recalculate the ID of the next image to be saved
A new button have been included in File|Properties form ("Settings" tab): "Recalc Next ID..."
It is enabled only when storage mode is not EmbeddedRTF and file has no modifications

Can be useful when we have inserted many images in some nodes of a file and later decide to move that nodes
(with all the images) to another file, so removing the nodes and images from the source file.
(To move the nodes with the images to a another file can be done exporting to a knt file format, for example)

This way the counter of next ID will be adjusted to the last available image ID in the file.

It looks for the max ID in the image list (so ignoring deleted images).

If configured Next ID > 1 + Max ID then it will show the following message:
 "Max ID in image list is %d and Next ID is %d
  Do you want the NEXT image to be saved with ID = %d
  * YOU MUST MAKE SURE there are no images with larger IDs on the external storage,
  perhaps referenced by other knt files (New images could override existing files)
  CONTINUE? "

If Next ID = 1 + Max ID it will show the following message:
  "Current Next ID (%d) cannot be reduced
  (Max ID in image list is %d)"