Modified the implementation of "Fix scrollbars" (e.g. to use with Linux-Wine)
Based on the behavior described in "Linux Wine Compatibility Notes Jan 1st 2024 + Request to see if text area scrollbar glitch can be fixed" (#633),
especially in the comment in https://github.com/dpradov/keynote-nf/issues/633#issuecomment-2016696660, it seems that a possible solution to this
problem could be to ensure that the RichEdit control has focus before starting its update (loading the content of the selected node)

I have reused the existing option (Fix scrollbars), available in Configuration Options...|Advanced (or in the .ini file: [KeyOptions] FixScrollBars)
for which the existing implementation does not seem to be helping.

On this occasion what is ensured is that at the time of loading the editor (TKntFolder.DataStreamToEditor) the editor has the focus, establishing
it if it is not (whenever possible), and returning the focus to the control that had it at the end . All this ensuring that the above does not
trigger the Editor.OnEnter event.

Among other things, and unlike the previous implementation, this one should not involve any flickering.

Note: At the moment I have not verified that it actually solves the problem. If the problem continues to persist, I will look for another solution.