If FirstTimeRun show the help file instead of trying to show "sample.knt"
If INI file was not found, we're probably being used 1st time after installation.
If no .KNT was specified, instead of showing the sample file which currenlty is not part of the distribution,
show the help file.
The first time, do not show the hint dialog and do not automatically create a new file.
Don't try to check for updates either.