* Adjustment in "Normal vs maximum tree width" functionality
The way in which the width of the tree is restored when the focus is passed from it to the RTF editor by clicking
with the mouse has been modified.
Depending on the speed of the press, text could be selected in an unwanted way. Also, the partial solution that
sought to avoid this was causing the selection made by following a bookmark (with selection) or a Find All result
to not show the target text selected.