Fixed: When using ImgFormatInsideRTF = ifWmetafile8 (.INI option), images in editor were removed when executing "Restore Image[s] proportions"
As indicated in help file:
<<
In order to make the images visible in the editor, in RTF, the formats currently available in the RichEdit control that can be used are: \wmetafile8, \emfblip, \pngblip or \jpegblip
When inserting the image programmatically, the format that KeyNote NF will used depends on this option:
-	ifWmetafile8 [0]
All images will be embedded in RTF with the format \wmetafile8. It will be the only option offered on systems with an older version of RichEdit (<= 4 ).
-	ifAccordingImage [1]
The application will use jpegblip for JPG images, emfblip for images that come from EMF files, wmetafile8 for those that come from WMF files, and pngblip for the rest: PNG, GIF, TIF, BMP and ICO.
>>

ifAccordingImage allows to use a more optimum formats and the size of images will be greatly reduced, and it is the default value in KeyNote.