Fixed: Macro 'xxx' not found [Reload Resource panel data when displaying]
If the "Reload resource panel data on screen" option was selected, when displaying the resource panel and selecting a tab
other than Macro, the macro list was cleared and therefore the macro execution was seen affected.
This option now allows you to reload the list of items for the selected tab in the resources panel from disk, but without
keeping them cleared.